requirements.txt
requests
bs4
spacy
newspaper3k